details drawer for transactions and users
