+++ db.json
@@ -6,7 +6,7 @@
       "category": "expenses",
       "receiver": "emag",
       "price": "600",
-      "VAT": "10",
+      "VAT": "12",
       "total_price": 0,
       "products": [
         {
@@ -108,7 +108,7 @@
       "date": "05.18.2021",
       "category": "expenses",
       "receiver": "Elefant",
-      "price": "1000",
+      "price": "777",
       "VAT": 10,
       "total_price": 0,
       "products": [
@@ -1049,7 +1049,7 @@
   "users": [
     {
       "first_name": "Andrei",
-      "last_name": "Mihu",
+      "last_name": "Mih",
       "id": 1005,
       "email": "Mihai_Mihu@gmail.com",
       "password": "rtyfghvbn",
@@ -1059,7 +1059,7 @@
     },
     {
       "first_name": "Anastasia",
-      "last_name": "Mirel",
+      "last_name": "Mirela",
       "id": 1006,
       "email": "Anastasia_Mirela@gmail.com",
       "password": "yuihjkbnm",
@@ -1069,7 +1069,7 @@
     },
     {
       "first_name": "Catalin",
-      "last_name": "Para",
+      "last_name": "Parara",
       "id": 1007,
       "email": "Catalin_Para@gmail.com",
       "password": "poilkjmnb",
@@ -1089,7 +1089,7 @@
     },
     {
       "first_name": "Darius",
-      "last_name": "Berenghi",
+      "last_name": "Bere",
       "id": 1010,
       "email": "Darius_Berenghi@gmail.com",
       "password": "789456123",
@@ -1108,7 +1108,7 @@
       "permissions": "Genin"
     },
     {
-      "first_name": "Andreea",
+      "first_name": "Andrei",
       "last_name": "Cretu",
       "id": 1012,
       "email": "Andreearetu@gmail.com",
@@ -1128,7 +1128,7 @@
       "permissions": "Chunin"
     },
     {
-      "first_name": "Valentina",
+      "first_name": "Valentin",
       "last_name": "Caprioreanu",
       "id": 1014,
       "email": "Valentin_Caprioreanu@gmail.com",
