+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Suspense, lazy, useState } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import { Provider, useDispatch } from 'react-redux';
+import { Provider } from 'react-redux';
 import store from "./context/redux/store";
 import DrawerMenu from "./components/DrawerMenu";
 import Header from "./components/Header";
@@ -9,14 +9,13 @@ import Footer from "./components/Footer";
 import ToastContext from "./context/ToastContext";
 import ToastNotification from "./components/Toast alerts/ToastNotification";
 import TransactionDetails from "./pages/transactions/TransactionDetails";
-import DetailsDrawer from "./components/DetailsDrawer";
+import DrawerDetails from "./components/DrawerDetails";
 
 const Transactions = lazy(() => import("./pages/transactions/Transactions"));
 const Payouts = lazy(() => import("./pages/payouts/Payouts"));
 const Users = lazy(() => import("./pages/users/Users"));
 const ErrorPage = lazy(() => import("./pages/ErrorPage"));
 const Dashboard = lazy(() => import("./pages/dashboard/Dashboard"));
-const UserDetails = lazy(() => import("./pages/users/UserDetails"));
 
 export default function App() {
   const { t } = useTranslation();
@@ -32,7 +31,7 @@ export default function App() {
         <>
           <Router>
             <ToastNotification />
-            <DetailsDrawer />
+            <DrawerDetails />
             <Header handleDrawerOpen={ handleDrawerOpen } open={ open } />
             <DrawerMenu handleDrawerClose={ handleDrawerClose } open={ open } />
             <Suspense fallback={ <div>{ t("loading_message") }</div> }>
@@ -40,10 +39,10 @@ export default function App() {
                 <Route path="/" element={ <Dashboard /> } />
                 <Route path="/dashboard" element={ <Dashboard /> } />
                 <Route path="/transactions" element={ <Transactions /> } />
-                <Route path="/transactions/:id" element={ <TransactionDetails /> } />
+                <Route path="/transactions/:id" element={ <Transactions /> } />
                 <Route path="/payouts" element={ <Payouts /> } />
                 <Route path="/users" element={ <Users /> } />
-                <Route path="/users/:id" element={ <UserDetails /> } />
+                <Route path="/users/:id" element={ <Users /> } />
                 <Route path="*" element={ <ErrorPage /> } />
               </Routes>
             </Suspense>
