+++ src/pages/transactions/TransactionValidationSchema.js
@@ -1,0 +1,10 @@
+import * as Yup from "yup";
+
+const TransactionValidationSchema = Yup.object({
+    receiver: Yup.string().required("required"),
+    price: Yup.number().required("required"),
+    VAT: Yup.number().required("required"),
+    products: Yup.array()
+});
+
+export default TransactionValidationSchema;
